:star-struck: Add option to go back to home page
